Complete auth settings refactor
