+++ routes/web.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\Front\DashboardController;
 use App\Http\Controllers\Front\AccountController;
 use App\Http\Controllers\Front\Settings\DeleteController;
 use App\Http\Controllers\Front\Settings\InformationController;
+use App\Http\Controllers\Front\Settings\SecurityController;
 use App\Http\Controllers\Front\WelcomeController;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
@@ -28,7 +29,8 @@ Route::group(['middleware' => ['auth', 'forbid-banned-user']], static function (
     Route::group(['prefix' => 'account-instellngen'], static function (): void {
         Route::get('/informatie', [InformationController::class, 'index'])->name('account.settings.information');
         Route::patch('/informatie', [InformationController::class, 'update'])->name('account.settings.information');
-        Route::patch('/security', [AccountController::class, 'updateSecurity'])->name('account.settings.security');
+        Route::get('/security', [SecurityController::class, 'index'])->name('account.settings.security');
+        Route::patch('/security', [SecurityController::class, 'update'])->name('account.settings.security');
     });
 
    Route::get('/home', DashboardController::class)->name('home');
